% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_api_mainland_company.R
\name{lxr_cn_company}
\alias{lxr_cn_company}
\alias{lxr_cn_company_equity_change}
\alias{lxr_cn_company_candlestick}
\alias{lxr_cn_company_dividend}
\alias{lxr_cn_company_fs_non_financial}
\alias{lxr_cn_company_fs_bank}
\alias{lxr_cn_company_fs_security}
\alias{lxr_cn_company_fs_insurance}
\alias{lxr_cn_company_fs_other_financial}
\title{Mainland company API endpoints}
\usage{
lxr_cn_company()

lxr_cn_company_equity_change()

lxr_cn_company_candlestick()

lxr_cn_company_dividend()

lxr_cn_company_fs_non_financial()

lxr_cn_company_fs_bank()

lxr_cn_company_fs_security()

lxr_cn_company_fs_insurance()

lxr_cn_company_fs_other_financial()
}
\value{
A string of the API endpoint.
}
\description{
These functions return the API endpoints for different aspects of a Mainland
company.
}
