% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_check_params.R
\name{lxr_check_params}
\alias{lxr_check_params}
\title{Check Query Parameters}
\usage{
lxr_check_params(query_params, valid_params, required_params)
}
\arguments{
\item{query_params}{A named list of query parameters.}

\item{valid_params}{A character vector of valid parameter names.}

\item{required_params}{A character vector of required parameter names.}
}
\value{
This function does not return a value. If the check is successful,
the function will complete silently. If there are invalid or missing
parameters, the function will stop and return an error message.
}
\description{
This function checks if the provided query parameters are valid and complete.
It compares the names of the query parameters with the valid and required
parameters for the specified API endpoint. If there are invalid or missing
parameters, it will stop the function and return an error message.
}
