% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_cn_company.R
\name{lxr_cn_company}
\alias{lxr_cn_company}
\title{Query Lixinger Chinese Company Data}
\usage{
lxr_cn_company(
  token = NULL,
  fs_type = NULL,
  mutual_markets = NULL,
  stock_codes = NULL,
  include_delisted = NULL
)
}
\arguments{
\item{token}{Mandatory API token for authentication.}

\item{fs_type}{Type of financial statements to query (optional). Options:
'non_financial', 'bank', 'insurance', 'security', 'other_financial'.}

\item{mutual_markets}{Markets to include in the query (optional). Options: 'ha'.}

\item{stock_codes}{Vector of stock codes to query (optional).}

\item{include_delisted}{Boolean indicating whether to include delisted
companies in the results (optional, defaults to FALSE).}
}
\value{
A tibble with the requested company data, with each market provided by
        mutualMarkets unnested into longer format.
}
\description{
Retrieves company data from the Lixinger API for Chinese companies. It allows
for specification of financial statement types, markets, stock codes, and the
option to include delisted companies in the results.
}
\examples{
\dontrun{
lxr_cn_company(
  token = "your_api_token",
  stock_codes = c("300750", "600519", "600157")
  )
}

}
\references{
For more detailed information about the parameters and usage of the Lixinger API
that this function interfaces with, please visit the Lixinger API documentation:
\url{https://www.lixinger.com/open/api/doc?api-key=cn/company}
}
