% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_cn_company_candlestick.R
\name{lxr_cn_company_candlestick}
\alias{lxr_cn_company_candlestick}
\title{Lixinger Chinese Company Candlestick Data}
\usage{
lxr_cn_company_candlestick(
  token = Sys.getenv("TOKEN_LIXINGER"),
  type,
  start_date = NULL,
  end_date = NULL,
  adjust_forward_date = NULL,
  adjust_backward_date = NULL,
  limit = NULL,
  stock_code
)
}
\arguments{
\item{token}{A string specifying the authentication token for the API.}

\item{type}{A string defining the type of candlestick data to retrieve.
The possible values are 'ex_rights', 'lxr_fc_rights', 'fc_rights', and
'bc_rights'.}

\item{start_date}{A Date object indicating the start date for data retrieval.
Dates should be in 'YYYY-MM-DD' format.The time span between `start_date`
and `end_date` must not exceed 10 years.}

\item{end_date}{End date for the query in "YYYY-MM-DD" format. If not provided,
the API defaults to the last Monday.}

\item{adjust_forward_date}{A string representing the date in "YYYY-MM-DD" format
for forward adjustment of the data range.}

\item{adjust_backward_date}{A string representing the date in "YYYY-MM-DD" format
for backward adjustment of the data range.}

\item{limit}{An integer specifying the maximum number of records to retrieve.}

\item{stock_code}{A string representing the stock code to query.}
}
\value{
A tibble (tbl_df) containing the flattened candlestick data.
}
\description{
This function queries the Lixinger API for candlestick data of Chinese companies.
}
\examples{
result <- lxr_cn_company_candlestick(
  type = "lxr_fc_rights",
  start_date = as.Date("2023-01-03"),
  stock_code = "300750"
)

}
