% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_cn_company_dividend.R
\name{lxr_cn_company_dividend}
\alias{lxr_cn_company_dividend}
\title{Query Chinese Company Dividend Data}
\usage{
lxr_cn_company_dividend(
  token = Sys.getenv("TOKEN_LIXINGER"),
  start_date,
  end_date = NULL,
  limit = NULL,
  stock_code
)
}
\arguments{
\item{token}{A character string. The user's unique API token.}

\item{start_date}{A character string in the format "YYYY-MM-DD". The start
date of the date range for which to retrieve data. The time span between
\code{start_date} and \code{end_date} must not exceed 10 years.}

\item{end_date}{A character string in the format "YYYY-MM-DD". The end
date of the date range for which to retrieve data.and the default value is
the date of the last Monday.}

\item{limit}{An integer. The maximum number of recent data records to
return. It only takes effect when the requested data is a date range.}

\item{stock_code}{A character string. The stock code of the company.
Please refer to the Stock Information API for valid stock codes.}
}
\value{
A tibble. The company's dividend data within the specified
date range.
}
\description{
This function queries the Lixinger API for dividend data of Chinese
companies.
}
\examples{
lxr_cn_company_dividend(
  start_date = "2020-01-01",
  end_date = "2024-12-31",
  stock_code = "300750"
)

}
\references{
For more detailed information about the parameters and usage of the Lixinger
API that this function interfaces with, please visit the Lixinger
\href{https://www.lixinger.com/open/api/doc?api-key=cn/company/dividend}{Doc}.
}
