% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_cn_company_equity_change.R
\name{lxr_cn_company_equity_change}
\alias{lxr_cn_company_equity_change}
\title{Query Equity Change Data for Chinese Companies}
\usage{
lxr_cn_company_equity_change(
  token = NULL,
  startDate = NULL,
  endDate = NULL,
  limit = NULL,
  stockCode = NULL
)
}
\arguments{
\item{token}{API token for authentication (string).}

\item{startDate}{Start date for the query in "YYYY-MM-DD" format (string).}

\item{endDate}{Optional end date for the query in "YYYY-MM-DD" format. If
not provided, the API defaults to the last Monday (string). The time span
between `startDate` and `endDate` must not exceed 10 years.}

\item{limit}{Optional integer specifying the maximum number of records to
return.}

\item{stockCode}{The stock code of the company (string).}
}
\value{
A tibble with flattened API response data.
}
\description{
Retrieves changes in equity from the Lixinger API for listed Chinese
companies.
}
\examples{
lxr_cn_company_equity_change(startDate = "2022-09-20",
                             limit = 20,
                             stockCode = "002352")
}
