% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_cn_company_fs.R
\name{lxr_cn_company_fs}
\alias{lxr_cn_company_fs}
\title{Query Financial Statements}
\usage{
lxr_cn_company_fs(
  token = Sys.getenv("TOKEN_LIXINGER"),
  fs_type = "non_financial",
  date = NULL,
  start_date = NULL,
  end_date = NULL,
  stock_codes,
  metrics_list
)
}
\arguments{
\item{token}{Character string, API access authentication token.}

\item{fs_type}{Character string, type of financial statements, see Details.}

\item{date}{Character string, specific date or "latest" in "YYYY-MM-DD" format.}

\item{start_date}{Character string, start date in "YYYY-MM-DD" format.}

\item{end_date}{Character string, end date in "YYYY-MM-DD" format.}

\item{stock_codes}{Character vector, stock codes for data retrieval (max 100).}

\item{metrics_list}{Character vector, metrics to retrieve.}
}
\value{
A tibble of the flattened financial data.
}
\description{
Fetches financial statement data for Chinese companies from Lixinger API.
}
\details{
fs_type should be one of the following: 'bank', 'insurance',
'security', 'non_financial', 'reit', 'other_financial'.
}
\examples{
# Fetch data for a specific date
lxr_cn_company_fs(
  date = "2023-09-30",
  stock_codes = c("300750", "600519", "600157"),
  metrics_list = c("q.ps.toi.t")
)

# Fetch data over a date range
lxr_cn_company_fs(
  start_date = "2022-09-30",
  end_date = "2023-09-30",
  stock_codes = "300750",
  metrics_list = "q.ps.toi.t"
)
}
\references{
For more detailed information about the parameters and usage of the Lixinger API
that this function interfaces with, please visit the Lixinger API documentation:
\url{https://www.lixinger.com/open/api/doc?api-key=cn/company/fs/non_financial}
}
