% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint-name.R
\name{lxr_endpoint_name}
\alias{lxr_endpoint_name}
\title{Extract the endpoint name}
\usage{
lxr_endpoint_name(endpoint)
}
\arguments{
\item{endpoint}{A string of the Lixinger endpoint URL.}
}
\value{
A string of the endpoint name.
}
\description{
This function takes a lixinger endpoint URL as input. It first extracts the
part of the URL after "api/". Then, all slashes ("/") and hyphens ("-") in
this part are replaced with underscores ("_"). The result is returned as the
endpoint name.
}
