% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_query.R
\name{lxr_query}
\alias{lxr_query}
\title{Lixinger API Query}
\usage{
lxr_query(
  url = NULL,
  token = NULL,
  fs_type = NULL,
  mutual_markets = NULL,
  stock_codes = NULL,
  include_delisted = NULL,
  timeout = 9,
  max_tries = 5
)
}
\arguments{
\item{url}{API endpoint URL for the Lixinger service.}

\item{token}{Mandatory API authentication token.}

\item{fs_type}{Optional type of financial statements to retrieve ('non_financial',
'bank', 'insurance', 'security', 'other_financial').}

\item{mutual_markets}{Optional markets to include, e.g., 'ha'.}

\item{stock_codes}{Optional stock codes to query, as a vector.}

\item{include_delisted}{Optional flag to include delisted stocks.}

\item{timeout}{Optional API request timeout in seconds, defaults to 9.}

\item{max_tries}{Optional number of retry attempts, defaults to 5.}
}
\value{
A list structure containing the Lixinger API response in a parsed JSON format.
}
\description{
Sends a request to the Lixinger API to retrieve financial data based on user-
specified parameters, such as financial statement type, market, stock codes,
and whether to include delisted stocks.
}
\note{
An API token is required to access Lixinger services. A valid API endpoint
      URL is also needed. Providing an invalid `token` or `url` will cause an error.
}
\examples{
\dontrun{
lxr_query(url = "https://open.lixinger.com/api/cn/company",
          token = "your_api_token",
          stock_codes = c("300750", "600519", "600157"))
}

}
