% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_query_params.R
\name{lxr_query_params}
\alias{lxr_query_params}
\title{Generate query parameters for Lixinger API}
\usage{
lxr_query_params(endpoint)
}
\arguments{
\item{endpoint}{A function representing the Lixinger API.}
}
\value{
A list containing two elements: 'valid_params' and 'required_params'.
'valid_params' is a vector of strings representing the names of the valid
parameters for the API. 'required_params' is a vector of strings representing
the names of the required parameters for the API.
}
\description{
This function generates a list of valid and required parameters for the
specified Lixinger API. The parameters are determined based on the API function.
}
\examples{
lxr_query_params(endpoint = lxr_cn_company())

}
