% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{lxr_unnest}
\alias{lxr_unnest}
\title{Unnest dataframes in list columns}
\usage{
lxr_unnest(df)
}
\arguments{
\item{df}{A dataframe with one or more list columns containing nested
dataframes to be unnested.}
}
\value{
A dataframe where each list column has been unnested into
multiple rows.
}
\description{
Unnest a dataframe with list columns, where the list columns
contain nested dataframes.
}
