% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_cn_company.R
\name{lxr_cn_company}
\alias{lxr_cn_company}
\title{Query Lixinger Chinese Company Data}
\usage{
lxr_cn_company(
  token = NULL,
  fsType = NULL,
  mutualMarkets = NULL,
  stockCodes = NULL,
  includeDelisted = NULL
)
}
\arguments{
\item{token}{API token for authentication.}

\item{fsType}{(Optional) Type of financial statements to query. Options:
'non_financial', 'bank', 'insurance', 'security', 'other_financial'.}

\item{mutualMarkets}{(Optional) Markets to include in the query. Options: 'ha'.}

\item{stockCodes}{(Optional) Vector of stock codes to query.}

\item{includeDelisted}{(Optional) Boolean indicating whether to include delisted
companies in the results.}
}
\value{
A tibble with the requested company data, with each market provided by
        mutualMarkets unnested into longer format.
}
\description{
Retrieves company data from the Lixinger API for Chinese companies. It allows
for specification of financial statement types, markets, stock codes, and the
option to include delisted companies in the results.
}
\examples{
# Retrieve information for all stocks
lxr_cn_company()

# Retrieve information for all stocks with a specific type of financial statements
lxr_cn_company(fsType = "non_financial")

# Retrieve information for specific stocks
lxr_cn_company(stockCodes = c("300750", "600519", "600157"))

# Retrieve data for stocks in the Shanghai-Hong Kong Stock Connect (Shanghai segment)
lxr_cn_company(mutualMarkets = "ha")
}
\references{
For more detailed information about the parameters and usage of the Lixinger API
that this function interfaces with, please visit the Lixinger API documentation:
\url{https://www.lixinger.com/open/api/doc?api-key=cn/company}
}
