% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_cn_company.R
\name{lxr_cn_company}
\alias{lxr_cn_company}
\alias{lxr_cn_company_equity_change}
\alias{lxr_cn_company_candlestick}
\alias{lxr_cn_company_shareholders_num}
\alias{lxr_cn_company_senior_executive_shares_change}
\alias{lxr_cn_company_block_deal}
\alias{lxr_cn_company_pledge}
\alias{lxr_cn_company_operation_revenue_constitution}
\alias{lxr_cn_company_operating_data}
\alias{lxr_cn_company_dividend}
\alias{lxr_cn_company_fs_non_financial}
\alias{lxr_cn_company_fs_bank}
\alias{lxr_cn_company_fs_security}
\alias{lxr_cn_company_fs_insurance}
\alias{lxr_cn_company_fs_other_financial}
\alias{lxr_cn_company_indices}
\alias{lxr_cn_company_industries}
\title{China Mainland company API endpoints}
\usage{
lxr_cn_company()

lxr_cn_company_equity_change()

lxr_cn_company_candlestick()

lxr_cn_company_shareholders_num()

lxr_cn_company_senior_executive_shares_change()

lxr_cn_company_block_deal()

lxr_cn_company_pledge()

lxr_cn_company_operation_revenue_constitution()

lxr_cn_company_operating_data()

lxr_cn_company_dividend()

lxr_cn_company_fs_non_financial()

lxr_cn_company_fs_bank()

lxr_cn_company_fs_security()

lxr_cn_company_fs_insurance()

lxr_cn_company_fs_other_financial()

lxr_cn_company_indices()

lxr_cn_company_industries()
}
\value{
A string of the API endpoint.
}
\description{
These functions return the API endpoints for different aspects of a Mainland
company.
}
