% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn-company.R
\name{lxr_cn_company}
\alias{lxr_cn_company}
\alias{lxr_cn_company_equity_change}
\alias{lxr_cn_company_candlestick}
\alias{lxr_cn_company_shareholders_num}
\alias{lxr_cn_company_senior_executive_shares_change}
\alias{lxr_cn_company_block_deal}
\alias{lxr_cn_company_pledge}
\alias{lxr_cn_company_operation_revenue_constitution}
\alias{lxr_cn_company_operating_data}
\alias{lxr_cn_company_dividend}
\alias{lxr_cn_company_allotment}
\alias{lxr_cn_company_fs_non_financial}
\alias{lxr_cn_company_fs_bank}
\alias{lxr_cn_company_fs_security}
\alias{lxr_cn_company_fs_insurance}
\alias{lxr_cn_company_fs_other_financial}
\alias{lxr_cn_company_indices}
\alias{lxr_cn_company_industries}
\alias{lxr_cn_company_fundamental_statistics}
\alias{lxr_cn_company_trading_abnormal}
\alias{lxr_cn_company_announcement}
\alias{lxr_cn_company_measures}
\alias{lxr_cn_company_inquiry}
\alias{lxr_cn_company_majority_shareholders}
\alias{lxr_cn_company_nolimit_shareholders}
\alias{lxr_cn_company_fund_shareholders}
\alias{lxr_cn_company_fund_collection_shareholders}
\alias{lxr_cn_company_customers}
\alias{lxr_cn_company_suppliers}
\alias{lxr_cn_company_fundamental_non_financial}
\alias{lxr_cn_company_fundamental_bank}
\alias{lxr_cn_company_fundamental_security}
\title{China Mainland company API endpoints}
\usage{
lxr_cn_company()

lxr_cn_company_equity_change()

lxr_cn_company_candlestick()

lxr_cn_company_shareholders_num()

lxr_cn_company_senior_executive_shares_change()

lxr_cn_company_block_deal()

lxr_cn_company_pledge()

lxr_cn_company_operation_revenue_constitution()

lxr_cn_company_operating_data()

lxr_cn_company_dividend()

lxr_cn_company_allotment()

lxr_cn_company_fs_non_financial()

lxr_cn_company_fs_bank()

lxr_cn_company_fs_security()

lxr_cn_company_fs_insurance()

lxr_cn_company_fs_other_financial()

lxr_cn_company_indices()

lxr_cn_company_industries()

lxr_cn_company_fundamental_statistics()

lxr_cn_company_trading_abnormal()

lxr_cn_company_announcement()

lxr_cn_company_measures()

lxr_cn_company_inquiry()

lxr_cn_company_majority_shareholders()

lxr_cn_company_nolimit_shareholders()

lxr_cn_company_fund_shareholders()

lxr_cn_company_fund_collection_shareholders()

lxr_cn_company_customers()

lxr_cn_company_suppliers()

lxr_cn_company_fundamental_non_financial()

lxr_cn_company_fundamental_bank()

lxr_cn_company_fundamental_security()
}
\value{
A string of the API endpoint.
}
\description{
These functions return the API endpoints for different aspects of a Mainland
company.
\itemize{
\item \code{lxr_cn_company()} gets stock detailed information.
\item \code{lxr_cn_company_equity_change()} gets equity change data.
\item \code{lxr_cn_company_candlestick()} gets stock K line data.
\item \code{lxr_cn_company_shareholders_num()} gets number of shareholders data.
\item \code{lxr_cn_company_senior_executive_shares_change()} gets senior executive
shares change data.
\item \code{lxr_cn_company_block_deal()} gets block deal data.
\item \code{lxr_cn_company_pledge()} gets equity pledge data.
\item \code{lxr_cn_company_operation_revenue_constitution()} gets operation revenue
constitution data.
\item \code{lxr_cn_company_operating_data()} gets operating data information.
\item \code{lxr_cn_company_dividend()} gets dividend information.
\item \code{lxr_cn_company_fs_non_financial()} gets non-financial listed company
financial data.
\item \code{lxr_cn_company_fs_bank()} gets bank listed company financial data.
\item \code{lxr_cn_company_fs_security()} gets security listed company financial data.
\item \code{lxr_cn_company_fs_insurance()} gets insurance listed company financial data.
\item \code{lxr_cn_company_fs_other_financial()} gets other financial listed company
financial data.
\item \code{lxr_cn_company_indices()} gets stock indices information.
\item \code{lxr_cn_company_industries()} gets stock industry information.
\item \code{lxr_cn_company_fundamental_statistics()} gets stock fundamental
statistics data.
\item \code{lxr_cn_company_trading_abnormal()} gets trading abnormal data.
\item \code{lxr_cn_company_announcement()} gets announcement data.
\item \code{lxr_cn_company_measures()} gets regulatory measures information.
\item \code{lxr_cn_company_inquiry()} gets inquiry information.
\item \code{lxr_cn_company_majority_shareholders()} gets information on the top ten
shareholders' holdings.
\item \code{lxr_cn_company_nolimit_shareholders()} gets the holding
information of the top ten circulating shareholders.
\item \code{lxr_cn_company_fund_shareholders()} gets publicly offered fund
shareholding information.
\item \code{lxr_cn_company_fund_collection_shareholders()} gets fund company's
shareholding information.
\item \code{lxr_cn_company_allotment()} gets allotment information.
\item \code{lxr_cn_company_customers()} gets customer information.
\item \code{lxr_cn_company_suppliers()} gets supplier information.
\item \code{lxr_cn_company_fundamental_non_financial()} gets fundamental data for
non-financial companies, such as PE, PB, etc.
\item \code{lxr_cn_company_fundamental_bank()} gets fundamental data for bank, such as
PE, PB, etc.
\item \code{lxr_cn_company_fundamental_security()} gets fundamental data for
securities, such as PE, PB, etc.
}
}
