% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_cn_company_equity_change.R
\name{lxr_cn_company_equity_change}
\alias{lxr_cn_company_equity_change}
\title{Query Equity Change Data for Chinese Companies}
\usage{
lxr_cn_company_equity_change(
  token = Sys.getenv("TOKEN_LIXINGER"),
  start_date,
  end_date = NULL,
  limit = NULL,
  stock_code
)
}
\arguments{
\item{token}{API token for authentication (string).}

\item{start_date}{Start date for the query in "YYYY-MM-DD" format (string).}

\item{end_date}{End date for the query in "YYYY-MM-DD" format. If
not provided, the API defaults to the last Monday (string). The time span
between \code{start_date} and \code{end_date} must not exceed 10 years.}

\item{limit}{Integer specifying the maximum number of records to
return.}

\item{stock_code}{The stock code of the company (string).}
}
\value{
A tibble with flattened API response data.
}
\description{
Retrieves changes in equity from the Lixinger API for listed Chinese
companies.
}
\examples{
lxr_cn_company_equity_change(
  start_date = "2022-09-20",
  limit = 20,
  stock_code = "002352"
)

}
\references{
For more detailed information about the parameters and usage of the Lixinger API
that this function interfaces with, please visit the Lixinger
\href{https://www.lixinger.com/open/api/doc?api-key=cn/company/equity-change}{API documentation}.
}
