% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_cn_company_fs.R
\name{lxr_cn_company_fs}
\alias{lxr_cn_company_fs}
\title{Query Financial Statements}
\usage{
lxr_cn_company_fs(
  token = NULL,
  fsType = "non_financial",
  date = NULL,
  startDate = NULL,
  endDate = NULL,
  stockCodes = NULL,
  metricsList = NULL
)
}
\arguments{
\item{token}{Character string, API access authentication token.}

\item{fsType}{(Optional) Character string, type of financial statements, see Details.}

\item{date}{(Optional) Character string, specific date or "latest" in "YYYY-MM-DD" format.}

\item{startDate}{(Optional) Character string, start date in "YYYY-MM-DD" format.}

\item{endDate}{(Optional) Character string, end date in "YYYY-MM-DD" format.}

\item{stockCodes}{Character vector, stock codes for data retrieval (max 100).}

\item{metricsList}{Character vector, metrics to retrieve.}
}
\value{
A tibble of the flattened financial data.
}
\description{
Fetches financial statement data for Chinese companies from Lixinger API.
}
\details{
fs_type should be one of the following: 'bank', 'insurance',
'security', 'non_financial', 'reit', 'other_financial'.
}
\examples{
# Fetch data for a specific date
lxr_cn_company_fs(
  date = "2023-09-30",
  stockCodes = c("300750", "600519", "600157"),
  metricsList = c("q.ps.toi.t")
)

# Fetch data over a date range
lxr_cn_company_fs(
  startDate = "2022-09-30",
  endDate = "2023-09-30",
  stockCodes = "300750",
  metricsList = "q.ps.toi.t"
)
}
\references{
For more detailed information about the parameters and usage of the Lixinger API
that this function interfaces with, please visit the Lixinger API documentation:
\url{https://www.lixinger.com/open/api/doc?api-key=cn/company/fs/non_financial}
}
