% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_cn_company_fs.R
\name{lxr_cn_company_fs}
\alias{lxr_cn_company_fs}
\title{Query Financial Statements}
\usage{
lxr_cn_company_fs(
  token = NULL,
  fs_type = "non_financial",
  date = NULL,
  start_date = NULL,
  end_date = NULL,
  stock_codes = NULL,
  metrics = NULL
)
}
\arguments{
\item{token}{(Required) Character string, API access authentication token.}

\item{fs_type}{(Optional) Character string, type of financial statements, see Details.}

\item{date}{(Optional) Date or character string, specific date or "latest" in "YYYY-MM-DD".}

\item{start_date}{(Optional) Date or character string, start date in "YYYY-MM-DD".}

\item{end_date}{(Optional) Date or character string, end date in "YYYY-MM-DD".}

\item{stock_codes}{(Required) Character vector, stock codes for data retrieval (max 100).}

\item{metrics}{(Required) Character vector, metrics to retrieve, as 'metricsList' in API.}
}
\value{
A tibble of the flattened financial data.
}
\description{
Fetches financial statement data for Chinese companies from Lixinger API.
}
\details{
fs_type should be one of the following: 'bank', 'insurance',
'security', 'non_financial', 'reit', 'other_financial'.
}
\examples{
\dontrun{
lxr_cn_company_fs(token = "your_token",
                  fs_type = "non_financial",
                  date = "latest",
                  start_date = "2020-01-01",
                  end_date = "2020-12-31",
                  stock_codes = c("300750", "600519"),
                  metrics = c("q.ps.to.it", "q.bs.ar.c.y2"))
}
}
\references{
For more detailed information about the parameters and usage of the Lixinger API
that this function interfaces with, please visit the Lixinger API documentation:
\url{https://www.lixinger.com/open/api/doc?api-key=cn/company/fs/non_financial}
}
