% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_endpoint_name.R
\name{lxr_endpoint_name}
\alias{lxr_endpoint_name}
\title{Extract the API name from a Lixinger API URL}
\usage{
lxr_endpoint_name(endpoint)
}
\arguments{
\item{endpoint}{A string of the Lixinger API URL.}
}
\value{
A string of the API name.
}
\description{
This function takes a Lixinger API URL as input, and returns the API name,
which is the part of the URL after "api/". All slashes ("/") in the API name
are replaced with underscores ("_").
}
