% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-company.R
\name{lxr_cn_company}
\alias{lxr_cn_company}
\title{China Hong Kong company API endpoints}
\usage{
lxr_cn_company()
}
\value{
A string of the API endpoint.
}
\description{
These functions return the API endpoints for different aspects of a China
Hong Kong company.
\itemize{
\item \code{lxr_hk_company()} retrieve stock details.
}
}
