% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-company.R
\name{lxr_hk_company}
\alias{lxr_hk_company}
\alias{lxr_hk_company_candlestick}
\alias{lxr_hk_company_employee}
\alias{lxr_hk_company_repurchase}
\alias{lxr_hk_company_short_selling}
\alias{lxr_hk_company_operation_revenue_constitution}
\alias{lxr_hk_company_indices}
\alias{lxr_hk_company_industries}
\alias{lxr_hk_company_fundamental_statistics}
\alias{lxr_hk_company_announcement}
\alias{lxr_hk_company_fund_shareholders}
\alias{lxr_hk_company_fund_collection_shareholders}
\title{Hong Kong company API endpoints}
\usage{
lxr_hk_company()

lxr_hk_company_candlestick()

lxr_hk_company_employee()

lxr_hk_company_repurchase()

lxr_hk_company_short_selling()

lxr_hk_company_operation_revenue_constitution()

lxr_hk_company_indices()

lxr_hk_company_industries()

lxr_hk_company_fundamental_statistics()

lxr_hk_company_announcement()

lxr_hk_company_fund_shareholders()

lxr_hk_company_fund_collection_shareholders()
}
\value{
A string of the API endpoint.
}
\description{
These functions return the API endpoints for different aspects of a China
Hong Kong company.
\itemize{
\item \code{lxr_hk_company()} retrieve stock details.
\item \code{lxr_hk_company_candlestick()} obtain candlestick data.The ex-rights
calculation is only applied to the prices for the selected time period;
trading volume is not adjusted for ex-rights.
\item \code{lxr_hk_company_employee()} obtain employee data.
\item \code{lxr_hk_company_repurchase()} obtain repurchase data. Calculate the share
capital as the total of H shares.
\item \code{lxr_hk_company_short_selling()} retrieve short-selling data, with the
capital stock calculated being the total H shares.
\item \code{lxr_hk_company_operation_revenue_constitution()} obtain revenue
composition data.
\item \code{lxr_hk_company_indices()} get information on the stock index it belongs to.
\item \code{lxr_hk_company_industries()} get information on the industry the stock
belongs to.
\item \code{lxr_hk_company_fundamental_statistics()} get fundamental analysis
statistical data.
\item \code{lxr_hk_company_announcement()} get the announcement information.
\item \code{lxr_hk_company_fund_shareholders()} obtain information on domestic fund
holdings.
\item \code{lxr_hk_company_fund_collection_shareholders()} obtain the shareholding
information of domestic fund companies."
}
}
