% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk-company.R
\name{lxr_hk_company}
\alias{lxr_hk_company}
\alias{lxr_hk_company_candlestick}
\alias{lxr_hk_company_employee}
\alias{lxr_hk_company_repurchase}
\alias{lxr_hk_company_short_selling}
\alias{lxr_hk_company_operation_revenue_constitution}
\alias{lxr_hk_company_indices}
\alias{lxr_hk_company_industries}
\alias{lxr_hk_company_fundamental_statistics}
\alias{lxr_hk_company_announcement}
\alias{lxr_hk_company_fund_shareholders}
\alias{lxr_hk_company_fund_collection_shareholders}
\alias{lxr_hk_company_dividend}
\alias{lxr_hk_company_split}
\alias{lxr_hk_company_allotment}
\alias{lxr_hk_company_fundamental_non_financial}
\alias{lxr_hk_company_fundamental_bank}
\alias{lxr_hk_company_fundamental_security}
\alias{lxr_hk_company_fundamental_insurance}
\alias{lxr_hk_company_fundamental_reit}
\alias{lxr_hk_company_fs_non_financial}
\alias{lxr_hk_company_fs_bank}
\alias{lxr_hk_company_fs_security}
\alias{lxr_hk_company_fs_insurance}
\alias{lxr_hk_company_fs_reit}
\alias{lxr_hk_company_fs_other_financial}
\alias{lxr_hk_company_hot_tr_dri}
\title{Hong Kong company API endpoints}
\usage{
lxr_hk_company()

lxr_hk_company_candlestick()

lxr_hk_company_employee()

lxr_hk_company_repurchase()

lxr_hk_company_short_selling()

lxr_hk_company_operation_revenue_constitution()

lxr_hk_company_indices()

lxr_hk_company_industries()

lxr_hk_company_fundamental_statistics()

lxr_hk_company_announcement()

lxr_hk_company_fund_shareholders()

lxr_hk_company_fund_collection_shareholders()

lxr_hk_company_dividend()

lxr_hk_company_split()

lxr_hk_company_allotment()

lxr_hk_company_fundamental_non_financial()

lxr_hk_company_fundamental_bank()

lxr_hk_company_fundamental_security()

lxr_hk_company_fundamental_insurance()

lxr_hk_company_fundamental_reit()

lxr_hk_company_fs_non_financial()

lxr_hk_company_fs_bank()

lxr_hk_company_fs_security()

lxr_hk_company_fs_insurance()

lxr_hk_company_fs_reit()

lxr_hk_company_fs_other_financial()

lxr_hk_company_hot_tr_dri()
}
\value{
A string of the API endpoint.
}
\description{
These functions return the API endpoints for different aspects of a China
Hong Kong company.
\itemize{
\item \code{lxr_hk_company()} retrieve stock details.
\item \code{lxr_hk_company_candlestick()} obtain candlestick data.The ex-rights
calculation is only applied to the prices for the selected time period;
trading volume is not adjusted for ex-rights.
\item \code{lxr_hk_company_employee()} obtain employee data.
\item \code{lxr_hk_company_repurchase()} obtain repurchase data. Calculate the share
capital as the total of H shares.
\item \code{lxr_hk_company_short_selling()} retrieve short-selling data, with the
capital stock calculated being the total H shares.
\item \code{lxr_hk_company_operation_revenue_constitution()} obtain revenue
composition data.
\item \code{lxr_hk_company_indices()} get information on the stock index it belongs to.
\item \code{lxr_hk_company_industries()} get information on the industry the stock
belongs to.
\item \code{lxr_hk_company_fundamental_statistics()} get fundamental analysis
statistical data.
\item \code{lxr_hk_company_announcement()} get the announcement information.
\item \code{lxr_hk_company_fund_shareholders()} obtain information on domestic fund
holdings.
\item \code{lxr_hk_company_fund_collection_shareholders()} obtain the shareholding
information of domestic fund companies."
\item \code{lxr_hk_company_dividend()} get dividend information.
\item \code{lxr_hk_company_split()} get split data.
\item \code{lxr_hk_company_allotment()} obtain rights issue information.
\item \code{lxr_hk_company_fundamental_non_financial()} gets fundamental data for
non-financial companies, such as PE, PB, etc.
\item \code{lxr_hk_company_fundamental_bank()} gets fundamental data for bank, such as
PE, PB, etc.
\item \code{lxr_hk_company_fundamental_security()} gets fundamental data for
securities, such as PE, PB, etc.
\item \code{lxr_hk_company_fundamental_insurance()} gets fundamental data for
insurance company, such as PE, PB, etc.
\item \code{lxr_hk_company_fundamental_reit()} obtain fundamental data of REITs, such
as PE, PB, etc.
\item \code{lxr_hk_company_fs_non_financial()} gets non-financial listed company
financial data.
\item \code{lxr_hk_company_fs_bank()} gets bank listed company financial data.
\item \code{lxr_hk_company_fs_security} gets security listed company financial data.
\item \code{lxr_hk_company_fs_insurance()} get insurance company financial data.
\item \code{lxr_hk_company_fs_reit()} obtain financial data of REITs.
\item \code{lxr_hk_company_fs_other_financial()} gets other financial listed company
financial data.
\item \code{lxr_hk_company_hot_tr_dri()} obtain dividend reinvestment yield data.
}
}
