% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lxr_query.R
\name{lxr_query}
\alias{lxr_query}
\title{Lixinger API Query}
\usage{
lxr_query(
  url,
  token = Sys.getenv("TOKEN_LIXINGER"),
  timeout = 9,
  max_tries = 5,
  ...
)
}
\arguments{
\item{url}{API endpoint URL for the Lixinger service.}

\item{token}{API authentication token. If not provided, the function will
attempt to use the 'TOKEN_LIXINGER' environment variable.}

\item{timeout}{Timeout for the request, in seconds. Defaults to 9
seconds.}

\item{max_tries}{Number of retry attempts in case of request failure, defaults
to 5 attempts.}

\item{...}{Additional parameters for the query.}
}
\value{
A list containing the Lixinger API response in a parsed JSON format.
}
\description{
Sends a request to the Lixinger API to retrieve financial data.
}
\details{
The function constructs an HTTP request to the specified Lixinger API endpoint.
It supports automatic retries and timeout settings. The function also handles
the conversion of additional query parameters to the required JSON format.
}
